services/driver-service/pyproject.toml
[project]
name = "driver-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.0.0",
    "fastapi",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.48",
    "uvicorn",
]
[dependency-groups]
dev = ["pytest>=8.0"]

[tool.pytest.ini_options]
pythonpath = ["src"]
